@@ -1,7 +1,4 @@
-﻿using Dapper;
-using Microsoft.CSharp.RuntimeBinder;
-using OPS.Core.Dapper;
-using System;
+﻿using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Data;
@@ -9,6 +6,8 @@ using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
 using System.Text;
+using Microsoft.CSharp.RuntimeBinder;
+using OPS.Core.Dapper;
 
 namespace Dapper
 {
