@@ -1,10 +1,10 @@
-﻿using OPS.Core.DTO;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Data.Common;
 using System.Globalization;
 using System.Text;
 using System.Threading.Tasks;
+using OPS.Core.DTO;
 
 namespace OPS.Core.DAO
 {
