@@ -1,18 +1,20 @@
 ﻿using System;
 using System.Threading;
-using System.Threading.Tasks;
 using RestSharp;
 using Serilog;
 
 namespace OPS.Core
 {
     public static class RestClientExtension
-    {       
-        public static IRestResponse ExecuteWithAutoRetry(this RestClient client, IRestRequest request, int totalRetries = 5)
+    {
+        public static RestResponse GetWithAutoRetry(this RestClient client, RestRequest request, int totalRetries = 5)
         {
+            var watch = System.Diagnostics.Stopwatch.StartNew();
+            Log.Verbose("Chamando URL: {Url}", request.Resource);
+
             try
             {
-                var response = client.Execute(request);
+                var response = client.Get(request);
                 if (!string.IsNullOrEmpty(response.Content))
                     return response;
 
@@ -28,14 +30,19 @@ namespace OPS.Core
 
                 throw;
             }
+            finally
+            {
+                watch.Stop();
+                Log.Verbose("Requisição processada em {TimeElapsed:c}", watch.Elapsed);
+            }
         }
 
-        private static IRestResponse ReTry(RestClient client, IRestRequest request, int totalRetries)
+        private static RestResponse ReTry(RestClient client, RestRequest request, int totalRetries)
         {
             Log.Verbose("Tentativa {Tentativa} para a url {Url}", totalRetries, request.Resource);
 
             Thread.Sleep(TimeSpan.FromMinutes(1));
-            return ExecuteWithAutoRetry(client, request, totalRetries - 1);
+            return GetWithAutoRetry(client, request, totalRetries - 1);
         }
     }
 }
