@@ -1,12 +1,9 @@
-﻿using CsvHelper;
-using OPS.Core;
-using System;
-using System.Collections.Generic;
+﻿using System;
 using System.Globalization;
 using System.IO;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
+using CsvHelper;
+using OPS.Core;
 
 namespace OPS.Importador
 {
